# fix #8991 and enhance collection performance