--- a/changelog/9125.improvement.rst
+++ b/changelog/9125.improvement.rst
@@ -1,0 +1,1 @@
+Improve collection performance by caching the internal hook proxy objects.