@elbehery95 2021-09-29
I am getting different results when trying to import `HookCaller` It seems that `HookCaller` is defined inside [`pluggy.hooks._HookCaller` in v0.13.1](https://github.com/pytest-dev/pluggy/blob/b1b07a309653a42a5311810f71d0addd09a4ec81/src/pluggy/hooks.py#L201) and as [`pluggy._hooks._HookCaller` in v1.0.0](https://github.com/pytest-dev/pluggy/blob/4259fdd799d64f953631e47ddcb69c3074a15c13/src/pluggy/_hooks.py#L184)

`setup.cfg` defines pluggy dependency `pluggy>=0.12,<2.0`  in such case this dependency is no longer valid right ? or maybe i am importing the wrong `_HookCaller`