@RonnyPfannschmidt 2021-09-30
@elbehery95 indeed, its a "private class" its ok to select based on where its available for now, we should make a issue report to pluggy for making it (or a protocol for it) avaliable as type)