@elbehery95 2021-09-30
Added the missing `_HookCaller` type annotation. Is there is any reason for limiting `_getconftestmodules` cache size to 128 results. I noticed some extra collection speed-up when `maxsize=None`