@bluetech 2021-09-30
(Not a review yet, just a comment)

I'm trying to understand the pre-caching line profile you've shown in the PR description. Do I understand correctly that 64.3543 s is spent in `gethookproxy`, and 60.4% of that time is just in the line `Path(fspath)`?