@elbehery95 2021-09-30
> (Not a review yet, just a comment)
> 
> I'm trying to understand the pre-caching line profile you've shown in the PR description. Do I understand correctly that 64.3543 s is spent in `gethookproxy`, and 60.4% of that time is just in the line `Path(fspath)`?

Probably yes; this is the base implementation with different format
 
```python
Total time: 65.7615 s
File: pytest/src/_pytest/main.py
Function: gethookproxy at line 543

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   543                                               @profile
   544                                               def gethookproxy(self, fspath: "os.PathLike[str]"):
   545                                                   # Check if we have the common case of running
   546                                                   # hooks with all conftest.py files.
   547    205108     853594.0      4.2      1.3          pm = self.config.pluginmanager
   548    205108   39292528.0    191.6     59.8          p = Path(fspath)
   549    205108    2775737.0     13.5      4.2          import_mode = self.config.getoption("importmode")
   550    205108    1111320.0      5.4      1.7          root_path = self.config.rootpath
   551    205108   14185275.0     69.2     21.6          my_conftestmodules = pm._getconftestmodules(p, import_mode, rootpath=root_path)
   552    205108    1414525.0      6.9      2.2          remove_mods = pm._conftest_plugins.difference(my_conftestmodules)
   553    205108     703070.0      3.4      1.1          if remove_mods:
   554                                                       # One or more conftests are not in use at this fspath.
   555    205096    2740815.0     13.4      4.2              from .config.compat import PathAwareHookProxy
   556
   557    205096    2007707.0      9.8      3.1              proxy = PathAwareHookProxy(FSHookProxy(pm, remove_mods))
   558                                                   else:
   559                                                       # All plugins are active for this fspath.
   560        12         47.0      3.9      0.0              proxy = self.config.hook
   561    205108     676901.0      3.3      1.0          return proxy
```

Other usage for `pathlib` that consume time can be found in `_getconftestmodules`

```python
Timer unit: 1e-06 s

Total time: 11.6957 s
File: pytest/src/_pytest/config/__init__.py
Function: _getconftestmodules at line 525

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   525                                               @lru_cache(maxsize=128)
   526                                               @profile
   527                                               def _getconftestmodules(
   528                                                   self, path: Path, importmode: Union[str, ImportMode], rootpath: Path   529                                               ) -> List[types.ModuleType]:
   530      2242       9181.0      4.1      0.1          if self._noconftest:
   531                                                       return []
   532
   533      2242      90943.0     40.6      0.8          if path.is_file():
   534      1957      47316.0     24.2      0.4              directory = path.parent
   535                                                   else:
   536       285        911.0      3.2      0.0              directory = path
   537
   538                                                   # XXX these days we may rather want to use config.rootpath
   539                                                   # and allow users to opt into looking into the rootdir parent
   540                                                   # directories instead of requiring to specify confcutdir.
   541      2242       6647.0      3.0      0.1          clist = []
   542     23086     575779.0     24.9      4.9          for parent in reversed((directory, *directory.parents)):
   543     20844    3180991.0    152.6     27.2              if self._confcutdir and parent in self._confcutdir.parents:
   544      8968      27306.0      3.0      0.2                  continue
   545     11876     880322.0     74.1      7.5              conftestpath = parent / "conftest.py"
   546     11876     562783.0     47.4      4.8              if conftestpath.is_file():
   547      4169    6242670.0   1497.4     53.4                  mod = self._importconftest(conftestpath, importmode, rootpath)
   548      4169      16294.0      3.9      0.1                  clist.append(mod)
   549      2242      48137.0     21.5      0.4          self._dirpath2confmods[directory] = clist
   550      2242       6397.0      2.9      0.1          return clist
```
where looking up if parent in parents takes about 30% from the total time ; I keep thinking that using `is_relative_to` or equivalent can be faster but not sure if this is valid or not

```python
>>> from pathlib import Path
>>> a = Path() ; b = Path()/'src/config/__init__.py'
>>> timeit.timeit(lambda: a in b.parents, number=30000)
0.8800809980020858
>>> timeit.timeit(lambda: str(b).startswith(str(a)), number=30000)
0.06638047900923993
```