@RonnyPfannschmidt 2021-09-30
@elbehery95 great find, i believe it may be sensible to check the parts attributes for quick checking