@bluetech 2021-10-01
@elbehery95 I've made some changes in `main` branch that would affect the numbers here. Could you rebase and see how the PR behaves after?

Sorry for the hassle. The reason I'm hesitant to add caches is that they have memory overhead and invalidation issues and generally make the code harder to understand and refactor. Especially in this case which deals with imports and side effects. So I always try to optimize the underlying operation first.